refactor(config): move server-config tests into config::server
The `outbound_http` and `allow_exec_activities` tests exercise
`ServerConfigToml` / `AllowExecActivities`, so they belong in a `tests`
module in config/server.rs rather than in the deployment test module.
